Add check if server is in ignore list

+++ app/events/client/commands/deleteAllMessages.py
@@ -2,6 +2,7 @@ import asyncio
 from discord.ext import commands
 from app.vars.client import client
 from app.helpers import Notify
+from app.filesystem import ignore
 
 @client.command(aliases=['removeallmessages', 'DAM', 'clearChannel'])
 @commands.guild_only()
@@ -9,6 +10,11 @@ from app.helpers import Notify
 async def deleteAllMessages(ctx):
     notify = Notify(ctx=ctx, title = 'Deleting All Channel Messages...')
     notify.prepair()
+
+    if str(ctx.guild.id) in ignore.getIgnore():
+        notify.error(content='The server {} is being ignored'.format(ctx.guild.name))
+        return
+
     async for message in ctx.message.channel.history():
         try:
             await message.delete()
